Add .gitlab-ci.yml file
To configure the CI/CD pipeline, this yaml file is added. This is only the template of the CI/CD file, so it needs to be updated for automated testing and static analysis. We can update the static analysis in this yaml file first, since automated testing needs to wait until the actual code is done.